Separate dev warnings tests. Issue platform-support warning once only.
